More Card components added. Card-specific css moved to specific component file.
