+++ src/components/cards/CardHeading.jsx
@@ -1,4 +1,11 @@
-import { StyledHeadingCard } from "./Cards";
+import styled from '@emotion/styled';
+import { BaseCard } from "./Cards";
+
+export const StyledHeadingCard = styled(BaseCard)`
+    h3 {
+        margin: 0;
+    }
+`
 
 export function HeadingCard({ children }) {
     return (
