+++ src/index.css
@@ -74,18 +74,27 @@ body {
   text-align: center;
 }
 
-h1, h2, h3, h4, h5, h6 {
+h1, h2 {
   font-family: var(--font-heading);
-  font-weight: 500;
   color: var(--text-h);
+}
+
+h3, h4, h5, h6 {
+  font-family: var(--font-heading);
+  color: var(--text);
+
 }
 
 h1 {
   font-size: 3rem;
 }
 
-h2, h3 {
+h2 {
   font-size: 2rem;
+}
+
+h3 {
+  font-size: 1.5rem;
 }
 
 h5 {
